LayoutTests/accessibility/image-role-is-synonomous-with-img-role-expected.txt

This test ensures that role="image" and role="img" are synonyms.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS img.role is 'AXRole: AXImage'
PASS image.role is 'AXRole: AXImage'
PASS successfullyParsed is true

TEST COMPLETE
